DAT-1341-object-specific-relations

+++ cmdb/interface/rest_api/routes/user_management_routes/users_routes.py
@@ -63,17 +63,17 @@ users_blueprint = APIBlueprint('users', __name__)
 @verify_api_access(required_api_level=ApiLevel.SUPER_ADMIN)
 @users_blueprint.protect(auth=True, right='base.user-management.user.add')
 @users_blueprint.validate(CmdbUser.SCHEMA)
-def insert_user(data: dict, request_user: CmdbUser):
+def insert_cmdb_user(data: dict, request_user: CmdbUser):
     """
     HTTP `POST` route to insert a CmdbUser into the database
 
     Args:
-        `data` (CmdbUser.SCHEMA): Data of a new CmdbUser
+        data (CmdbUser.SCHEMA): Data of a new CmdbUser
 
     Returns:
-        `InsertSingleResponse`: Insert response with the new CmdbUser and the corresponding public_id
+        InsertSingleResponse: Insert response with the new CmdbUser and the corresponding public_id
     """
-    #TODO: REFATOR-FIX
+    #TODO: REFATOR-FIX (reduce complexity)
     try:
         users_manager: UsersManager = ManagerProvider.get_manager(ManagerType.USERS_MANAGER, request_user)
         security_manager: SecurityManager = ManagerProvider.get_manager(ManagerType.SECURITY_MANAGER, request_user)
@@ -89,16 +89,13 @@ def insert_user(data: dict, request_user: CmdbUser):
         except KeyError:
             return abort(400, "The database of the user could not be retrieved!")
 
-        try:
-            if current_app.cloud_mode:
-                # Confirm an email was provided when creating the user
-                user_email = data['email']
+        if current_app.cloud_mode:
+            # Confirm an email was provided when creating the user
+            user_email = data.get('email')
 
-                if not user_email:
-                    raise KeyError
-        except KeyError:
-            LOGGER.error("[insert_user] No email was provided!")
-            return abort(400, "The email is mandatory to create a new user!")
+            if not user_email:
+                LOGGER.error("[insert_user] No email was provided!")
+                return abort(400, "The email is mandatory to create a new user!")
 
         # Check if email already exists
         try:
@@ -108,7 +105,7 @@ def insert_user(data: dict, request_user: CmdbUser):
                 if user_with_given_email:
                     return abort(400, "The email is already in use!")
         except UsersManagerGetError:
-            pass
+            return abort(400, "Failed to retrieve User from database!")
 
         if current_app.cloud_mode and current_app.local_mode:
             # Open file and check if user exists
@@ -132,21 +129,23 @@ def insert_user(data: dict, request_user: CmdbUser):
         result_id = users_manager.insert_user(data)
 
         #Confirm that user is created
-        user = users_manager.get_user(result_id)
+        created_user = users_manager.get_user(result_id)
+        if created_user:
+            api_response = InsertSingleResponse(CmdbUser.to_json(created_user), result_id)
 
-        api_response = InsertSingleResponse(CmdbUser.to_dict(user), result_id)
+            return api_response.make_response()
 
-        return api_response.make_response()
+        return abort(404, "Could not retrieve the created User from the database!")
     except HTTPException as http_err:
         raise http_err
     except UsersManagerInsertError as err:
-        LOGGER.error("[insert_user] %s", err, exc_info=True)
-        return abort(400, "Could not create the user in database!")
+        LOGGER.error("[insert_cmdb_user] %s", err, exc_info=True)
+        return abort(400, "Failed to create the User in database!")
     except UsersManagerGetError as err:
-        LOGGER.error("[insert_user] %s", err, exc_info=True)
-        return abort(500, "Could not retrieve the created user from the database!")
+        LOGGER.error("[insert_cmdb_user] %s", err, exc_info=True)
+        return abort(500, "Failed to retrieve the created User from the database!")
     except Exception as err:
-        LOGGER.error("[insert_user] Exception: %s. Type: %s", err, type(err), exc_info=True)
+        LOGGER.error("[insert_cmdb_user] Exception: %s. Type: %s", err, type(err), exc_info=True)
         return abort(500, "Internal server error!")
 
 # ---------------------------------------------------- CRUD - READ --------------------------------------------------- #
@@ -156,15 +155,15 @@ def insert_user(data: dict, request_user: CmdbUser):
 @verify_api_access(required_api_level=ApiLevel.ADMIN)
 @users_blueprint.protect(auth=True, right='base.user-management.user.view')
 @users_blueprint.parse_collection_parameters()
-def get_users(params: CollectionParameters, request_user: CmdbUser):
+def get_cmdb_users(params: CollectionParameters, request_user: CmdbUser):
     """
     HTTP `GET`/`HEAD` route for retrieving multiple CmdbUsers with a filter
 
     Args:
-        `params` (CollectionParameters): Passed parameters over the http query string
+        params (CollectionParameters): Passed parameters over the http query string
 
     Returns:
-        `GetMultiResponse`: The CmdbUsers matching the given filter
+        GetMultiResponse: The CmdbUsers matching the given filter
     """
     try:
         users_manager: UsersManager = ManagerProvider.get_manager(ManagerType.USERS_MANAGER, request_user)
@@ -172,7 +171,7 @@ def get_users(params: CollectionParameters, request_user: CmdbUser):
         builder_params = BuilderParameters(**CollectionParameters.get_builder_params(params))
 
         iteration_result: IterationResult[CmdbUser] = users_manager.iterate(builder_params)
-        users = [CmdbUser.to_dict(user) for user in iteration_result.results]
+        users = [CmdbUser.to_json(user) for user in iteration_result.results]
 
         api_response = GetMultiResponse(users,
                                         total=iteration_result.total,
@@ -182,45 +181,46 @@ def get_users(params: CollectionParameters, request_user: CmdbUser):
 
         return api_response.make_response()
     except UsersManagerIterationError as err:
-        LOGGER.error("[get_users] %s", err, exc_info=True)
+        LOGGER.error("[get_cmdb_users] %s", err, exc_info=True)
         return abort(400, "Could not iterate the requested users!")
     except Exception as err:
-        LOGGER.error("[get_users] Exception: %s. Type: %s", err, type(err), exc_info=True)
+        LOGGER.error("[get_cmdb_users] Exception: %s. Type: %s", err, type(err), exc_info=True)
         return abort(500, "Internal server error!")
-
-
 
 
 @users_blueprint.route('/<int:public_id>', methods=['GET', 'HEAD'])
 @insert_request_user
 @verify_api_access(required_api_level=ApiLevel.ADMIN)
 @users_blueprint.protect(auth=True, right='base.user-management.user.view', excepted={'public_id': 'public_id'})
-def get_user(public_id: int, request_user: CmdbUser):
+def get_cmdb_user(public_id: int, request_user: CmdbUser):
     """
     HTTP `GET`/`HEAD` route for a single CmdbUser
 
     Args:
-        `public_id` (int): public_id of the requested CmdbUser
+        public_id (int): public_id of the requested CmdbUser
 
     Returns:
-        `GetSingleResponse`: Raw data of the requested CmdbUser
+        GetSingleResponse: Raw data of the requested CmdbUser
     """
     try:
         users_manager: UsersManager = ManagerProvider.get_manager(ManagerType.USERS_MANAGER, request_user)
 
-        user: CmdbUser = users_manager.get_user(public_id)
+        requested_user = users_manager.get_user(public_id)
 
-        api_response = GetSingleResponse(CmdbUser.to_dict(user), body=request.method == 'HEAD')
+        if requested_user:
+            api_response = GetSingleResponse(CmdbUser.to_json(requested_user), body=request.method == 'HEAD')
 
-        return api_response.make_response()
+            return api_response.make_response()
+
+        return abort(404, f"The User with ID:{public_id} was not found!")
+    except HTTPException as http_err:
+        raise http_err
     except UsersManagerGetError as err:
-        LOGGER.error("[get_user] %s", err, exc_info=True)
-        return abort(404, f"Could not retrieve the user with public_id: {public_id}")
+        LOGGER.error("[get_cmdb_user] %s", err, exc_info=True)
+        return abort(400, f"Failed to retrieve the User with ID: {public_id}!")
     except Exception as err:
-        LOGGER.error("[get_user] Exception: %s. Type: %s", err, type(err), exc_info=True)
+        LOGGER.error("[get_cmdb_user] Exception: %s. Type: %s", err, type(err), exc_info=True)
         return abort(500, "Internal server error!")
-
-
 
 # --------------------------------------------------- CRUD - UPDATE -------------------------------------------------- #
 
@@ -229,31 +229,39 @@ def get_user(public_id: int, request_user: CmdbUser):
 @verify_api_access(required_api_level=ApiLevel.SUPER_ADMIN)
 @users_blueprint.protect(auth=True, right='base.user-management.user.edit', excepted={'public_id': 'public_id'})
 @users_blueprint.validate(CmdbUser.SCHEMA)
-def update_user(public_id: int, data: dict, request_user: CmdbUser):
+def update_cmdb_user(public_id: int, data: dict, request_user: CmdbUser):
     """
     HTTP `PUT`/`PATCH` route to update a single CmdbUser
 
     Args:
-        `public_id` (int): public_id of the CmdbUser which should be updated
-        `data` (CmdbUser.SCHEMA): New values for the CmdbUser
+        public_id (int): public_id of the CmdbUser which should be updated
+        data (CmdbUser.SCHEMA): New values for the CmdbUser
 
     Returns:
-        `UpdateSingleResponse`: The updated raw data of the CmdbUser
+        UpdateSingleResponse: The updated raw data of the CmdbUser
     """
     try:
         users_manager: UsersManager = ManagerProvider.get_manager(ManagerType.USERS_MANAGER, request_user)
 
-        user = CmdbUser.from_data(data=data)
-        users_manager.update_user(public_id, user)
+        to_update_user = users_manager.get_user(public_id)
 
-        api_response = UpdateSingleResponse(CmdbUser.to_dict(user))
+        if to_update_user:
+            user = CmdbUser.from_data(data=data)
 
-        return api_response.make_response()
+            users_manager.update_user(public_id, user)
+
+            api_response = UpdateSingleResponse(CmdbUser.to_json(user))
+
+            return api_response.make_response()
+
+        return abort(404, f"The User with ID:{public_id} was not found!")
+    except HTTPException as http_err:
+        raise http_err
     except UsersManagerUpdateError as err:
-        LOGGER.error("[update_user] %s", err, exc_info=True)
-        return abort(400, f"Could not update the user with public_id: {public_id}!")
+        LOGGER.error("[update_cmdb_user] %s", err, exc_info=True)
+        return abort(400, f"Failed to update the User with public_id: {public_id}!")
     except Exception as err:
-        LOGGER.error("[update_user] Exception: %s. Type: %s", err, type(err), exc_info=True)
+        LOGGER.error("[update_cmdb_user] Exception: %s. Type: %s", err, type(err), exc_info=True)
         return abort(500, "Internal server error!")
 
 
@@ -261,39 +269,43 @@ def update_user(public_id: int, data: dict, request_user: CmdbUser):
 @insert_request_user
 @verify_api_access(required_api_level=ApiLevel.SUPER_ADMIN)
 @users_blueprint.protect(auth=True, right='base.user-management.user.edit', excepted={'public_id': 'public_id'})
-def change_user_password(public_id: int, request_user: CmdbUser):
+def change_cmdb_user_password(public_id: int, request_user: CmdbUser):
     """
     HTTP `PATCH` route for changing the password of a CmdbUser
 
     Args:
-        `public_id` (int): public_id of the CmdbUser
+        public_id (int): public_id of the CmdbUser
 
     Returns:
-        `UpdateSingleResponse`:  The CmdbUser with new password
+        UpdateSingleResponse:  The CmdbUser with new password
     """
     try:
         users_manager: UsersManager = ManagerProvider.get_manager(ManagerType.USERS_MANAGER, request_user)
         security_manager: SecurityManager = ManagerProvider.get_manager(ManagerType.SECURITY_MANAGER, request_user)
 
-        user = users_manager.get_user(public_id)
+        to_update_user = users_manager.get_user(public_id)
 
-        password = security_manager.generate_hmac(request.json.get('password'))
-        user.password = password
-        users_manager.update_user(public_id, user)
+        if to_update_user:
+            password = security_manager.generate_hmac(request.json.get('password'))
+            to_update_user.password = password
+            users_manager.update_user(public_id, to_update_user)
 
-        api_response = UpdateSingleResponse(CmdbUser.to_dict(user))
+            api_response = UpdateSingleResponse(CmdbUser.to_json(to_update_user))
 
-        return api_response.make_response()
+            return api_response.make_response()
+
+        return abort(404, f"The User with ID:{public_id} was not found!")
+    except HTTPException as http_err:
+        raise http_err
     except UsersManagerGetError as err:
-        LOGGER.error("[change_user_password] %s", err, exc_info=True)
-        return abort(404, f"Could not retrieve the user with public_id: {public_id}")
+        LOGGER.error("[change_cmdb_user_password] %s", err, exc_info=True)
+        return abort(400, f"Failed to retrieve the User with ID: {public_id}")
     except UsersManagerUpdateError as err:
-        LOGGER.error("[update_user] %s", err, exc_info=True)
-        return abort(400, f"Could not change the password for user with public_id: {public_id}!")
+        LOGGER.error("[change_cmdb_user_password] %s", err, exc_info=True)
+        return abort(400, f"Failed to change the password for User with ID: {public_id}!")
     except Exception as err:
-        LOGGER.error("[update_user] Exception: %s. Type: %s", err, type(err), exc_info=True)
+        LOGGER.error("[change_cmdb_user_password] Exception: %s. Type: %s", err, type(err), exc_info=True)
         return abort(500, "Internal server error!")
-
 
 # --------------------------------------------------- CRUD - DELETE -------------------------------------------------- #
 
@@ -301,30 +313,36 @@ def change_user_password(public_id: int, request_user: CmdbUser):
 @insert_request_user
 @verify_api_access(required_api_level=ApiLevel.SUPER_ADMIN)
 @users_blueprint.protect(auth=True, right='base.user-management.user.delete')
-def delete_user(public_id: int, request_user: CmdbUser):
+def delete_cmdb_user(public_id: int, request_user: CmdbUser):
     """
     HTTP `DELETE` route to delete a single CmdbUser
 
     Args:
-        `public_id` (int): public_id of the CmdbUser
+        public_id (int): public_id of the CmdbUser
 
     Returns:
-        `DeleteSingleResponse`: Raw data of the deleted CmdbUser
+        DeleteSingleResponse: Raw data of the deleted CmdbUser
     """
     try:
         users_manager: UsersManager = ManagerProvider.get_manager(ManagerType.USERS_MANAGER, request_user)
 
-        deleted_group = users_manager.delete_user(public_id)
+        to_delete_user = users_manager.get_user(public_id)
 
-        api_response = DeleteSingleResponse(raw=CmdbUser.to_dict(deleted_group))
+        if to_delete_user:
+            users_manager.delete_user(public_id)
 
-        return api_response.make_response()
+            api_response = DeleteSingleResponse(raw=CmdbUser.to_json(to_delete_user))
+            return api_response.make_response()
+
+        return abort(404, f"The User with ID:{public_id} was not found!")
+    except HTTPException as http_err:
+        raise http_err
     except UsersManagerDeleteError as err:
-        LOGGER.error("[delete_user] %s", err, exc_info=True)
-        return abort(400, f"Could not delete user with ID: {public_id} !")
+        LOGGER.error("[delete_cmdb_user] %s", err, exc_info=True)
+        return abort(400, f"Failed to delete User with ID: {public_id}!")
     except UsersManagerGetError as err:
-        LOGGER.error("[delete_user] %s", err, exc_info=True)
-        return abort(404, f"Could not retrieve the user with ID: {public_id}")
+        LOGGER.error("[delete_cmdb_user] %s", err, exc_info=True)
+        return abort(404, f"Failed to retrieve the User with ID: {public_id}!")
     except Exception as err:
-        LOGGER.error("[delete_user] Exception: %s. Type: %s", err, type(err), exc_info=True)
+        LOGGER.error("[delete_cmdb_user] Exception: %s. Type: %s", err, type(err), exc_info=True)
         return abort(500, "Internal server error!")
